# openapi.yaml
openapi: 3.1.0
info:
  title: Sleeper Backend API
  description: |
    A comprehensive fantasy football API that aggregates data from **KeepTradeCut (KTC)** and **Sleeper API**.

    This API provides comprehensive fantasy football data by combining:
    - 🏈 **KeepTradeCut (KTC)**: Player rankings and trade values (Dynasty & Redraft rankings, 1QB & Superflex league formats, TEP scoring variations)
    - 🏈 **Sleeper API**: Player data & league management (Comprehensive player profiles, League management, Research data and projections)

    ## Performance Optimizations
    - **Database Caching Strategy**: First API call fetches from external APIs and caches in database, subsequent calls served instantly from local cache
    - **Response Time Guidelines**: First refresh call takes 30-60 seconds, cached data retrieval < 1 second, health checks < 100ms
    - **Gunicorn Multi-Worker Setup**: 9 worker processes for concurrent request handling with load balancing
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/rawarne2/sleeper-backend
servers:
  - url: http://localhost:5001
    description: Development server
  - url: https://sleeper-backend.vercel.app
    description: Production server
...
paths:
...
  /api/sleeper/league/{league_id}/stats/week/{week}:
    get:
      tags:
        - Sleeper Weekly Stats
      summary: Get weekly stats for a specific week
      description: |
        Get weekly fantasy points and roster information for all players in a specific week.
        Returns player scoring data including points, roster ID, and is_starter status.
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/WeekPath"
        - $ref: "#/components/parameters/SeasonQuery"
        - $ref: "#/components/parameters/LeagueType"
        - name: average
          in: query
          description: Whether to return season averages (weeks 1-17; week 18 excluded)
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Weekly stats retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyStatsResponse"
        "404":
          description: Weekly stats not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Sleeper Weekly Stats
      summary: Refresh weekly stats for a specific week
      description: |
        Fetch fresh weekly stats data from Sleeper API for a specific week and save to database.
        This endpoint fetches matchup data and extracts player scoring information.
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/WeekPath"
        - $ref: "#/components/parameters/SeasonQuery"
        - $ref: "#/components/parameters/LeagueType"
      responses:
        "200":
          description: Weekly stats refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyStatsRefreshResponse"
        "400":
          description: Invalid parameters or failed to refresh
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Sleeper Weekly Stats
      summary: Update weekly stats for a specific week
      description: |
        Fetch fresh weekly stats data from Sleeper API for a specific week and save to database.
        This endpoint fetches matchup data and extracts player scoring information.
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/WeekPath"
        - $ref: "#/components/parameters/SeasonQuery"
        - $ref: "#/components/parameters/LeagueType"
      responses:
        "200":
          description: Weekly stats refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyStatsRefreshResponse"
        "400":
          description: Invalid parameters or failed to refresh
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/sleeper/league/{league_id}/stats/seed:
    post:
      tags:
        - Sleeper Weekly Stats
      summary: Seed league stats information
      description: |
        Create or update league information needed for weekly stats tracking.
        This should be called before fetching weekly stats to ensure proper league setup.
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                league_name:
                  type: string
                  description: The league name
                  default: "Fantasy League"
                season:
                  type: string
                  description: The NFL season year
                  pattern: "^[0-9]{4}$"
                  example: "2024"
                league_type:
                  type: string
                  description: League type
                  enum: [ "dynasty", "redraft" ]
                  default: "dynasty"
                scoring_settings:
                  type: object
                  description: League scoring settings
              required:
                - league_name
                - season
      responses:
        "200":
          description: League stats seeded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueStatsSeedResponse"
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
...
    LeagueId:
      name: league_id
      in: path
      description: The Sleeper league ID
      required: true
      schema:
        type: string
        pattern: "^[0-9]+$"
      example: "1210364682523656192"
...
    SeasonQuery:
      name: season
      in: query
      description: The NFL season year
      required: false
      schema:
        type: string
        pattern: "^[0-9]{4}$"
        default: "2024"
      example: "2024"
    WeekPath:
      name: week
      in: path
      description: Week number (1-18 for regular season)
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 18
      example: 1
...
    LeagueType:
      name: league_type
      in: query
      description: League type - "dynasty" or "redraft"
      required: false
      schema:
        type: string
        enum: [ "dynasty", "redraft" ]
        default: "dynasty"
      example: "dynasty"
  schemas:
...
    ErrorResponse:
      type: object
      description: >
        Standard JSON error envelope. Optional fields depend on the route;
        clients should tolerate unknown keys.
      required:
        - status
        - error
        - timestamp
      additionalProperties: true
      properties:
        status:
          type: string
          enum: [ "error" ]
        error:
          type: string
          description: Error message
        details:
          type: string
          description: >
            Additional detail as text. When the source value is structured
            (e.g. upstream API error object), the server may JSON-encode it into this string.
        timestamp:
          type: string
          format: date-time
          description: Error timestamp (RFC 3339, UTC with Z)
        hint:
          type: string
          description: Operator hint (e.g. maintenance auth)
        suggestion:
          type: string
          description: Suggested remediation (e.g. which refresh to run)
        job_id:
          type: string
        league_id:
          type: string
        season:
          type: string
        week:
          oneOf:
            - type: integer
            - type: string
          description: Week number or the literal "all" for research refresh errors
        league_type:
          type: string
          enum: [ "dynasty", "redraft" ]
        parameters:
          type: object
          additionalProperties: true
          description: Request parameters associated with the error (e.g. KTC format flags)
        configuration:
          type: object
          additionalProperties: true
          description: Server-side configuration context (e.g. cleanup target)
        results:
          type: array
          description: Per-item outcomes (e.g. prewarm summary entries)
          items:
            type: object
            additionalProperties: true
        sleeper_players_fetched:
          type: integer
          description: Count from Sleeper merge failures
...
    WeeklyStatsResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ "success" ]
        data_type:
          type: string
          enum: [ "weekly", "averages" ]
        season:
          type: string
        week:
          type: integer
        weeks_considered:
          type: string
        records:
          type: array
          items:
            type: object
            properties:
              player_id:
                type: string
              roster_id:
                type: integer
              points:
                type: number
              is_starter:
                type: boolean
              last_updated:
                type: string
                format: date-time
        player_averages:
          type: object
        timestamp:
          type: string
          format: date-time
    WeeklyStatsRefreshResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ "success" ]
        message:
          type: string
          example: "Weekly stats refreshed successfully"
        league_id:
          type: string
        season:
          type: string
        week:
          type: integer
        refresh_results:
          type: object
        refreshed_stats:
          type: object
        timestamp:
          type: string
          format: date-time
    LeagueStatsSeedResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ "success" ]
        message:
          type: string
          example: "League stats seeded successfully"
        action:
          type: string
          enum: [ "created", "updated" ]
        league_id:
          type: string
        league_name:
          type: string
        season:
          type: string
        timestamp:
          type: string
          format: date-time
...